fixed #4
according to many internet threads, the warn log can be ignored and will
be removed in future versions. so the message is now filtered out.

added / (slash) to the and of all request pathes in the serverConsole.html,
because the webserver didn't interpret the path as the resource path
when the slash was missing. This was causing the "console in console"
behavior (f.e.: the request for the serverFriends didn't return the
serverFriends list, but the whole website again...)